move forwarddecl and includes to separate modules
