+++ crates/re_types/source_hash.txt
@@ -1,4 +1,4 @@
 # This is a sha256 hash for all direct and indirect dependencies of this crate's build script.
 # It can be safely removed at anytime to force the build script to run again.
 # Check out build.rs to see how it's computed.
-2ac93ae4b1911839224fbf4208ce13ca62e54ab9fef3884f777d3c16ca746a32
+7198e7f974640d28afd8771c7bc91cbafb7a4c2332a290c781aca00bb8836e5a
